Ignore mounted devices
Previously, if attempting to export a VM which has isos, or images mounted to it the code would attempt to export these mounts, which causes a failure attempting to locate the targetId. This causes the vmdk to not be downloaded at all and fail out.

Pull request merely does a check for targetId, implicitly checking whether or not the iterated device is exportable.