Minor fixes for content-length and http range
Changed the code to set Content-length always.
This required some changes for the tarfile to seek if
size property does not exist.

Also the range was 1 byte too much, so it would leave 1
byte left on the connection. I'm not sure if this is a
problem, but good to fix.