Added search by VM Name capability in upload_file_to_vm script
This is a fix for issue #443.
This script currently takes only the vm uuid to identify the VM and
upload a file to it.
Added a command line argument (i.e.) vm_name since using uuid is a
little cumbersome. This allows the users to specify vm_name OR vm_uuid.